Compiled debug and release for x86 and x64

+++ RayEngine/Source/DX11/DX11DepthStencilView.cpp
@@ -45,7 +45,7 @@ namespace RayEngine
 		/////////////////////////////////////////////////////////////
 		void DX11DepthStencilView::SetName(const std::string& name)
 		{
-			m_View->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str());
+			m_View->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(name.size()), name.c_str());
 		}
