+++ RayEngine/Source/DX11/DX11Factory.cpp
@@ -131,7 +131,7 @@ namespace RayEngine
 		/////////////////////////////////////////////////////////////
 		void DX11Factory::SetName(const std::string& name)
 		{
-			m_Factory->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str());
+			m_Factory->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(name.size()), name.c_str());
 		}
