+++ RayEngine/Source/DX11/DX11Sampler.cpp
@@ -63,7 +63,7 @@ namespace RayEngine
 		/////////////////////////////////////////////////////////////
 		void DX11Sampler::SetName(const std::string& name)
 		{
-			m_SamplerState->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str());
+			m_SamplerState->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(name.size()), name.c_str());
 		}
