+++ RayEngine/Source/DX11/DX11Shader.cpp
@@ -64,7 +64,7 @@ namespace RayEngine
 		/////////////////////////////////////////////////////////////
 		void DX11Shader::SetName(const std::string& name)
 		{
-			m_Shader->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str());
+			m_Shader->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(name.size()), name.c_str());
 		}
